Initial Tether Implementation
This is a simple one to start that will keep the player at the point that they attached themselves at and circle them around. On release of left mouse, the player is sent in the direction of their up vector at the default speed from the player controller. I will work on looking at speed increases after sending this in.

The tether is launched in the direction that the mouse points and will anchor the player only if the tether detects a gameobject in the AnchorPoint layer.